[Apps.] Regenerated demo for cdc_msd_sdcard_sam_a7g5_ek
MH3-93744

+++ apps/cdc_msd_sdcard/firmware/src/config/sam_a7g5_ek/driver/sdmmc/src/drv_sdmmc.c
@@ -1622,7 +1622,7 @@ static void DRV_SDMMC_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[GEOMETRY_TABLE_READ_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[GEOMETRY_TABLE_READ_ENTRY].numBlocks))
         {
             return;
         }
@@ -1633,7 +1633,7 @@ static void DRV_SDMMC_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
         {
             return;
         }
